% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiableR.r
\name{normEigenMS}
\alias{normEigenMS}
\title{EigenMS normalisation}
\usage{
normEigenMS(df, sample.info)
}
\arguments{
\item{df}{the tibble containing only FeatureIDs and sample abundance}

\item{sample.info}{sample info. Must contain a column named "sample"
for the sample information and a column named "group" for the experimental group}
}
\value{
a EigenMS normalised tibble dataframe
}
\description{
Use the ProteoMM package to normalise the dataframe using EigenMS
}
