% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiableR.r
\name{normMedian}
\alias{normMedian}
\title{Median normalisation}
\usage{
normMedian(df, mean = NA)
}
\arguments{
\item{df}{the tibble containing only FeatureIDs and sample abundance}

\item{mean}{the average value to normalise the dataframe to. Default as NA
which normalise the mean of the dataframe to the mean of the medians}
}
\value{
a Median normalised tibble dataframe
}
\description{
Median normalise the dataframe
}
