% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiableR.r
\name{runLimma}
\alias{runLimma}
\title{Limma differential analysis}
\usage{
runLimma(df, design.matrix, cont.matrix)
}
\arguments{
\item{df}{the tibble containing only FeatureIDs and sample abundance}

\item{design.matrix}{limma design matrix. See limma}

\item{cont.matrix}{limma cont matrix. See limma}
}
\value{
The log-fold change and p value of each feature per comparison
}
\description{
Perform limma differential analysis
}
