% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiableR.r
\name{tbManipulate}
\alias{tbManipulate}
\title{Data frame preprocesssing}
\usage{
tbManipulate(df, filter_threshold, impute = "gm2", log_base = 2)
}
\arguments{
\item{df}{the tibble containing only FeatureIDs and sample abundance}

\item{filter_threshold}{minimum number of non NAs to keep the feature}

\item{impute}{imputation method. "gm2" = half of global minimum, "gm5" = a fifth of global minimum}

\item{log_base}{log transform base. Default = 2}
}
\value{
processed and transformed tibble data frame
}
\description{
The function preprocess the data by filtering, imputing, and transforming
}
